fix dirs, added props without states and fix some styles
